Move newResidentAndResidencySchema
